Reset last_call_result after reading it
The root cause seems to be that two calls go through the whole
process step-by-step in lockstep. The the result of the second
overwrites last_call_result before it is read.

To properly detect such errors we reset last_call_result to null
after each time reading it, so that the same value cannot be read
twice.