Add delay into code to test
We are adding a delay by await async {} into the code to test.  This is
simulating as if the code did some other inter-canister call before doing the
call that we want to test. That is quite common.

It breaks the test. The Call and Release tester rely on the call arriving
before we can call release(). This was actually part of the reason why the
Stage tester was invented.