Stage tester reading from a variable
Now we are adding a second Stage tester where the `after` function reads from a
global variable. Now there is a timing problem. Because of the delay in fetch()
the call to run() arrives after the call to release(). This causes the after
method not to run at release() time but later at run() time. Now the timing is
off and after reads the wrong value.