Move after() from release() back into run()
Undoing a previous PR, we move the execution of the after method back in run()
where it was before. This allows us to call release() at any time and have
consistent behaviour of when the after method runs.

The ReleaseTester is rewritten with an internal Buffer to accomodate for the
change.

Two tests are removed that relied on `after` running at release-time. But they
can easily be replaced using the ReleaseTester.